rename Events.Misc to Events.Tags
