+++ iBot/Events/ChannelEventManager.cs
@@ -5,7 +5,7 @@ using IBot.Core;
 using IBot.Events.Args.RoomState;
 using IBot.Events.Args.UserList;
 using IBot.Events.Args.Users;
-using IBot.Events.Misc;
+using IBot.Events.Tags;
 
 namespace IBot.Events
 {
